# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AudiusSource": {
        "properties": {
          "provider": {
            "const": "audius",
            "title": "Provider",
            "type": "string"
          },
          "track_id": {
            "title": "Track Id",
            "type": "string"
          }
        },
        "required": [
          "provider",
          "track_id"
        ],
        "title": "AudiusSource",
        "type": "object"
      },
...
      "CreateAudiusJobRequest": {
        "properties": {
          "profile": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Profile"
          },
          "source": {
            "$ref": "#/components/schemas/AudiusSource"
          }
        },
        "required": [
          "source"
        ],
        "title": "CreateAudiusJobRequest",
        "type": "object"
      },
      "CreateObjectJobRequest": {
        "properties": {
          "input": {
            "$ref": "#/components/schemas/ObjectJobInput"
          },
          "profile": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Profile"
          }
        },
        "required": [
          "input"
        ],
        "title": "CreateObjectJobRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "JobArtifacts": {
        "properties": {
          "analysis": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Analysis",
            "type": "object"
          },
          "broad_stems": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Broad Stems",
            "type": "object"
          },
          "bundles": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Bundles",
            "type": "object"
          },
          "derived_stems": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Derived Stems",
            "type": "object"
          },
          "main_stems": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Main Stems",
            "type": "object"
          },
          "midi": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Midi",
            "type": "object"
          },
          "specialist_substems": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Specialist Substems",
            "type": "object"
          },
          "tempo_locked_wavs": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Tempo Locked Wavs",
            "type": "object"
          }
        },
        "title": "JobArtifacts",
        "type": "object"
      },
      "JobEventsResponse": {
        "properties": {
          "events": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Events",
            "type": "array"
          },
          "next_after": {
            "title": "Next After",
            "type": "integer"
          }
        },
        "required": [
          "events",
          "next_after"
        ],
        "title": "JobEventsResponse",
        "type": "object"
      },
      "JobResponse": {
        "additionalProperties": true,
        "properties": {
          "artifacts": {
            "$ref": "#/components/schemas/JobArtifacts"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "missing_features": {
            "items": {
              "type": "string"
            },
            "title": "Missing Features",
            "type": "array"
          },
          "rejected_candidates": {
            "additionalProperties": true,
            "title": "Rejected Candidates",
            "type": "object"
          },
          "remote_adapter_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remote Adapter Reason"
          },
          "remote_adapter_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remote Adapter Status"
          },
          "stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "timings": {
            "$ref": "#/components/schemas/JobTimings"
          }
        },
        "required": [
          "job_id",
          "status"
        ],
        "title": "JobResponse",
        "type": "object"
      },
      "JobTimings": {
        "additionalProperties": true,
        "properties": {
          "local_elapsed_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Local Elapsed Seconds"
          },
          "local_total_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Local Total Seconds"
          },
          "worker_total_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Worker Total Seconds"
          }
        },
        "title": "JobTimings",
        "type": "object"
      },
      "ObjectJobInput": {
        "properties": {
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "object": {
            "$ref": "#/components/schemas/ObjectReference"
          }
        },
        "required": [
          "filename",
          "object"
        ],
        "title": "ObjectJobInput",
        "type": "object"
      },
      "ObjectReference": {
        "additionalProperties": true,
        "properties": {
          "bucket": {
            "title": "Bucket",
            "type": "string"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          }
        },
        "required": [
          "provider",
          "bucket",
          "key"
        ],
        "title": "ObjectReference",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Asynchronous music source-separation control plane.",
    "title": "StemSplitter API",
    "version": "development"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/jobs": {
      "post": {
        "operationId": "createJob",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/CreateObjectJobRequest"
                  },
                  {
                    "$ref": "#/components/schemas/CreateAudiusJobRequest"
                  }
                ]
              }
            },
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "file": {
                    "format": "binary",
                    "type": "string"
                  },
                  "profile": {
                    "type": "string"
                  }
                },
                "required": [
                  "file"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Job Route",
        "tags": [
          "jobs"
        ]
      }
    },
    "/jobs/{job_id}": {
      "delete": {
        "operationId": "deleteJob",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Job Route",
        "tags": [
          "jobs"
        ]
      },
      "get": {
        "operationId": "getJob",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Job Status",
        "tags": [
          "jobs"
        ]
      }
    },
    "/jobs/{job_id}/cancel": {
      "post": {
        "operationId": "cancelJob",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Job Route",
        "tags": [
          "jobs"
        ]
      }
    },
    "/jobs/{job_id}/events": {
      "get": {
        "operationId": "getJobEvents",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "default": "0",
              "title": "After",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": "100",
              "title": "Limit",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobEventsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Job Events",
        "tags": [
          "jobs"
        ]
      }
    },
    "/jobs/{job_id}/manifest": {
      "get": {
        "operationId": "getJobManifest",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Job Manifest",
        "tags": [
          "jobs"
        ]
      }
    },
    "/jobs/{job_id}/resume": {
      "post": {
        "operationId": "resumeJob",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resume Job",
        "tags": [
          "jobs"
        ]
      }
    },
...
  }
}